Implemented the constant seeding pictures logic, added account profile pictures, as well

+++ WebPresentationLayer/Controllers/AccountController.cs
@@ -96,7 +96,7 @@ public class AccountController : Controller
                 var imageName = $"{Guid.NewGuid()}{extension}";
                 _fileSrv.SaveMemoryStreamToFile(fileBytes, "account", imageName);
                 //Delete old photo file
-                if (!String.IsNullOrWhiteSpace(dbUser.PhotoPath))
+                if (!String.IsNullOrWhiteSpace(dbUser.PhotoPath) && !dbUser.PhotoPath.Contains("Seeded"))
                     _fileSrv.DeleteFile(dbUser.PhotoPath);
 
                 dbUser.PhotoPath = $"/account/{imageName}";
@@ -288,7 +288,7 @@ public class AccountController : Controller
                 var imageName = $"{Guid.NewGuid()}{extension}";
                 _fileSrv.SaveMemoryStreamToFile(fileBytes, "pet", imageName);
                 //Delete old photo file
-                if (!String.IsNullOrWhiteSpace(dbPet.PhotoPath))
+                if (!String.IsNullOrWhiteSpace(dbPet.PhotoPath) && !dbPet.PhotoPath.Contains("Seeded"))
                     _fileSrv.DeleteFile(dbPet.PhotoPath);
 
                 dbPet.PhotoPath = $"/pet/{imageName}";
@@ -315,7 +315,7 @@ public class AccountController : Controller
         }
     }
 
-    [HttpGet("account/pets/create")]
+    [HttpGet("Account/PetCreate")]
     public async Task<IActionResult> PetCreate()
     {
         ViewBag.GenderOptions = Enum.GetValues(typeof(GenderEnum))
@@ -356,7 +356,6 @@ public class AccountController : Controller
                 //Save new photo file
                 var imageName = $"{Guid.NewGuid()}{extension}";
                 _fileSrv.SaveMemoryStreamToFile(fileBytes, "pet", imageName);
-                //Delete old photo file
 
                 newPet.PhotoPath = $"/pet/{imageName}";
             }
